requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
opencv-python>=4.7.0

# Data Science & Visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.22.0
notebook>=6.5.0

# Streamlit Dashboard
streamlit>=1.25.0

# Evaluation & Metrics
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0

# Optional: For better progress bars and visualizations
# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

